RavenDB-24310 Extract connection string logic to ClusterStateMachine.ConnectionStrings.cs partial
